Init: STEP 9 초기화
STEP8 에 대한 피드백 내용 추가 (README)
