--- Presenter/LogPresenter.js
@@ -1,14 +1,0 @@
-class VendingMachineLogPresenter {
-    constructor(model, view) {
-        this.model = model;
-    }
-
-    sendLogDataToModel(logData) {
-        this.model.insertLogData(logData);
-    }
-
-
-
-
-    
-}
