+++ Presenter/VendingMachineLogPresenter.js
@@ -1,0 +1,12 @@
+class VendingMachineLogPresenter {
+    constructor(model, view) {
+        this.model = model;
+    }
+
+    sendLogDataToModel(logData) {
+        this.model.insertLogData(logData);
+    }
+
+}
+
+export { VendingMachineLogPresenter }
