+++ main.html
@@ -231,7 +231,9 @@
         <div class="alert hidden center-tag middle"></div>
     </footer>
 
-    <script type="text/javascript" src="./util.js"></script>
+    <script type="module" src="http://localhost:8080/app.js"></script>
+
+    <!-- <script type="text/javascript" src="./util.js"></script>
     <script type="text/javascript" src="./app.js"></script>
     <script type="text/javascript" src="./Model/Model.js"></script>
     <script type="text/javascript" src="./View/MainView.js"></script>
@@ -243,7 +245,7 @@
     <script type="text/javascript" src="./Presenter/ItemPresenter.js"></script>
     <script type="text/javascript" src="./Presenter/ControlPresenter.js"></script>
     <script type="text/javascript" src="./Presenter/WalletPresenter.js"></script>
-    <script type="text/javascript" src="./Presenter/LogPresenter.js"></script>
+    <script type="text/javascript" src="./Presenter/LogPresenter.js"></script> -->
 
 </body>
 </html>
